+++ server/src/main/java/org/apache/iotdb/db/conf/IoTDBConfig.java
@@ -21,7 +21,7 @@ package org.apache.iotdb.db.conf;
 import org.apache.iotdb.common.rpc.thrift.TEndPoint;
 import org.apache.iotdb.commons.client.property.ClientPoolProperty.DefaultProperty;
 import org.apache.iotdb.commons.conf.IoTDBConstant;
-import org.apache.iotdb.commons.utils.PathUtils;
+import org.apache.iotdb.commons.utils.FileUtils;
 import org.apache.iotdb.commons.utils.TestOnly;
 import org.apache.iotdb.consensus.ConsensusFactory;
 import org.apache.iotdb.db.audit.AuditLogOperation;
@@ -1325,7 +1325,7 @@ public class IoTDBConfig {
     } catch (IOException e) {
       logger.error("Fail to get canonical path of {}", dataHomeFile, e);
     }
-    return PathUtils.addPrefix2FilePath(dataHomeDir, dir);
+    return FileUtils.addPrefix2FilePath(dataHomeDir, dir);
   }
 
   void confirmMultiDirStrategy() {
